requirements.txt
streamlit~=1.40.2
nltk~=3.9.1
numpy~=2.0.2
pymorphy2~=0.9.1
requests~=2.32.3
matplotlib~=3.9.3
natasha~=1.6.0
wordcloud~=1.9.4
bs4~=0.0.2
beautifulsoup4~=4.12.3
pillow~=11.0.0